+++ NppTranslatePanel/PluginInfrastructure/NppPluginNETHelper.cs
@@ -194,8 +194,15 @@ namespace Kbg.NppPluginNET.PluginInfrastructure
     [StructLayout(LayoutKind.Sequential)]
     public struct toolbarIcons
     {
-        public IntPtr hToolbarBmp;			// standard icon (color)
-        public IntPtr hToolbarIcon;			// Fluent UI icon (black)
-        public IntPtr hToolbarIconDarkMode;	// Fluent UI icon (white)
+		public IntPtr hToolbarBmp;			// standard icon (color)
+		public IntPtr hToolbarIcon;			// Fluent UI icon (black)
+	}
+
+    [StructLayout(LayoutKind.Sequential)]
+    public struct toolbarIconsWithDarkMode
+    {
+		public IntPtr hToolbarBmp;			// standard icon (color)
+		public IntPtr hToolbarIcon;			// Fluent UI icon (black)
+		public IntPtr hToolbarIconDarkMode;	// Fluent UI icon (white)
     }
 }