+++ NppTranslatePanel/PluginInfrastructure/UnmanagedExports.cs
@@ -50,6 +50,7 @@ namespace Kbg.NppPluginNET
             if (notification.Header.Code == (uint)NppMsg.NPPN_TBMODIFICATION)
             {
                 PluginBase._funcItems.RefreshItems();
+                Main.RegisterToolbarIcon();
             }
             else if (notification.Header.Code == (uint)NppMsg.NPPN_SHUTDOWN)
             {